Added SalonService Dtos and graphql schema

+++ src/main/java/com/verdant/salon_ecomm/dtos/reviews/ReviewConnection.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record ReviewConnection(
         List<ReviewDto> items,
-        long totalCount,
+        int totalCount,
         int totalPages,
         int currentPage
 ) {}
